Tamaño pelota ajustado

+++ PinballTemplate/Source/ModulePhysics.h
@@ -41,7 +41,7 @@ public:
 	update_status PostUpdate();
 	bool CleanUp();
 
-	PhysBody* CreateCircle(int x, int y, int radius);
+	PhysBody* CreateCircle(int x, int y, float radius);
 	PhysBody* CreateRectangle(int x, int y, int width, int height);
 	PhysBody* CreateChain(int x, int y, const int* points, int size);
 	PhysBody* CreateEdge(int x, int y, const int* points, int size);
